requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
joblib==1.4.2
numpy==2.2.1
scikit-learn==1.6.1
python-dotenv==1.0.1
supabase==2.10.0
pandas==2.2.3
httpx